creates expenses container and updates expenses page and component
